requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116

torch~=2.1.0
torchvision==0.14.1+cu116
faiss-cpu==1.7.3
pytorch-lightning==1.9.4
pytorch-metric-learning==2.0.1
opencv-python==4.7.0.72
scikit-image==0.19.3
gdown~=4.7.1

numpy~=1.26.1
Pillow~=10.1.0
scikit-learn~=1.3.1
tqdm~=4.66.1